+++ fh-v8/src/fh_sequel.js
@@ -1,0 +1,3 @@
+
+// actually call the prelude function, which calls the main() function implicitly
+prelude();
