+++ fh-v8/src/flow_heater.js
@@ -1,19 +1,4 @@
-
-
-async function dispatch_request(url, request) {
-    const spec = {
-        "url": url,
-        "request": request
-    };
-
-    await Deno.core.jsonOpAsync("dispatch_request", spec);
-}
-
-async function main() {
-    Deno.core.ops();
-
-    // run the get_request function (provided by the surrounding rust ecosystem)
-    let request = Deno.core.jsonOpSync("get_request", []);
+async function main(request) {
     Deno.core.print(`DENO: Got request body: ${request.body}, content-type header: ${request.headers['content-type']}, method: ${request.method}\n`);
 
     // modify the requests data and return it back to the rust-context
@@ -24,5 +9,3 @@ async function main() {
     await dispatch_request("http://httpbin.org/anything", request);
     await dispatch_request("http://httpbin.org/anything", request);
 }
-
-main();
