+++ fh-v8/src/lib.rs
@@ -4,7 +4,7 @@ mod runtime;
 
 use crate::request::{Request, RequestList};
 use crate::response::Response;
-use crate::runtime::prepare_runtime;
+use crate::runtime::{prepare_runtime, prepare_user_code};
 use anyhow::Result;
 use std::collections::HashMap;
 
@@ -12,9 +12,12 @@ pub async fn process_request(req: Request, code: Option<String>) -> Result<Respo
     let mut js_runtime = prepare_runtime(req.clone());
 
     if let Some(c) = code {
-        js_runtime.execute("custom_code.js", &c)?;
+        js_runtime.execute("custom_code.js", &prepare_user_code(&c, false))?;
     } else {
-        js_runtime.execute("flow_heater.js", include_str!("flow_heater.js"))?;
+        js_runtime.execute(
+            "flow_heater.js",
+            &prepare_user_code(include_str!("flow_heater.js"), true),
+        )?;
     }
 
     js_runtime.run_event_loop().await?;
